Potential Date Solution - Please Review
I created a class called ClassDate that can be added in a Vector or List
to the Student class.  Essentially all it does is store a date and
attendance status for that date.  I have included an edited version of
the Student class to use ClassDate, but I have not attempted to work
this in with serialization, especially since it sounds like we will be
going another route soon anyway.  Additionally, this allowed me to add a
getScore method to Student which calculates the grade that student has
for attendance based on the options found in my ScoringOptions class.