Add Import And Export to File Menu
I added and tested Import and Export options to the File dropdown in the
menu bar.  Had to make some changes to CSVPort to correct some
formatting problems.  I put the csv file that Leland gave me to test
with in the main directory if anyone wants to try it out or something.
When you import this file, it creates a new class and fills it with all
the students and their info.  The export function does the reverse, but
the scoring options are not yet implemented.  All scores are currently
returned as 100s.