Fixed AttdendanceDlg to have working combo boxes
The combo boxes can have the values "Present", "Absent", or "Tardy", and
they default to "Present" when the dialogue first pops up.  I attempted
to get attendance to update, but am getting a NullPointerException when
update the data store.  Needs investigation.  Most of the code needed to
update attendance is there though.